refactor(rspace): split rspace.rs into topical child modules
Mechanical modularization ahead of the RSpace/ReplayRSpace dedup: the
1344-line rspace.rs becomes a 99-line root (structs, module declarations,
accessors, SpaceMatcher marker, LOCK_SEQUENCE re-export) plus eight child
modules under rspace/:

- locks.rs        two-phase striped lock acquisition + LOCK_SEQUENCE
- setup.rs        apply/create/create_with_replay/create_history_repo/
                  spawn and hot-store rebuild
- ispace_impl.rs  the ISpace trait impl (entry points, checkpointing,
                  state accessors)
- trace_log.rs    event-log + produce-counter bookkeeping
- ops_consume.rs  locked consume path and its store helpers
- ops_produce.rs  locked produce path, matcher driver, candidate
                  retirement
- ops_install.rs  install registry and restore
- tests.rs        contention/diagnostic test module, moved verbatim

Every body moved unchanged; the only deltas are pub(super) visibility on
methods now called across sibling modules and per-file imports. Public
surface unchanged (LOCK_SEQUENCE re-exported at its old path). Full
rspace_plus_plus release suite green (12 binaries, 0 failed), clippy
clean, downstream rholang and casper compile unchanged.
