+++ rspace++/src/rspace/rspace.rs
@@ -3,45 +3,46 @@
 // NOTE: Manual marks are used instead of trace_i()/with_marks() because
 // the functions are not async-compatible with Span trait's closure pattern.
 // This matches Scala's Span[F].traceI() and withMarks() semantics.
+//
+// Module layout (child modules of this file):
+//   locks.rs       — two-phase striped lock acquisition + LOCK_SEQUENCE
+//   setup.rs       — construction/wiring (apply, create, create_with_replay,
+//                    create_history_repo, spawn, hot-store rebuild)
+//   ispace_impl.rs — the ISpace trait impl (operation entry points,
+//                    checkpointing, state accessors)
+//   trace_log.rs   — event-log + produce-counter bookkeeping
+//   ops_consume.rs — locked consume path and its store helpers
+//   ops_produce.rs — locked produce path, matcher driver, candidate
+//                    retirement
+//   ops_install.rs — install registry and restore
 
-use std::collections::{BTreeMap, BTreeSet, HashMap};
+use std::collections::{BTreeMap, HashMap};
 use std::fmt::Debug;
 use std::hash::Hash;
 use std::sync::Arc;
-use std::sync::atomic::{AtomicU64, Ordering as AtomicOrdering};
-use std::time::Instant;
 
-pub static LOCK_SEQUENCE: AtomicU64 = AtomicU64::new(0);
-
-use async_trait::async_trait;
-use rand::seq::SliceRandom;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use shared::rust::store::key_value_store::KeyValueStore;
 
-use super::checkpoint::SoftCheckpoint;
-use super::errors::{HistoryRepositoryError, RSpaceError};
-use super::hashing::blake2b256_hash::Blake2b256Hash;
-use super::history::history_reader::HistoryReader;
-use super::history::instances::radix_history::RadixHistory;
-use super::logging::BasicLogger;
-use super::r#match::Match;
-use super::metrics_constants::{
-    CHANGES_SPAN, CONSUME_COMM_LABEL, CREATE_CHECKPOINT_SPAN, HISTORY_CHECKPOINT_SPAN,
-    LOCKED_CONSUME_SPAN, LOCKED_PRODUCE_SPAN, PRODUCE_COMM_LABEL, RESET_SPAN,
-    REVERT_SOFT_CHECKPOINT_SPAN, RSPACE_METRICS_SOURCE,
-};
-use super::replay_rspace::ReplayRSpace;
-use super::rspace_interface::{
-    ContResult, ISpace, MaybeConsumeResult, MaybeProduceCandidate, MaybeProduceResult, RSpaceResult,
-};
-use super::striped_locks::{self, ChannelLockGuard};
-use super::trace::Log;
-use super::trace::event::{COMM, Consume, Event, IOEvent, Produce};
-use crate::rspace::checkpoint::Checkpoint;
-use crate::rspace::history::history_repository::{HistoryRepository, HistoryRepositoryInstances};
-use crate::rspace::hot_store::{HotStore, HotStoreInstances};
-use crate::rspace::internal::*;
+use crate::rspace::history::history_repository::HistoryRepository;
+use crate::rspace::hot_store::HotStore;
+use crate::rspace::internal::Install;
+use crate::rspace::r#match::Match;
 use crate::rspace::space_matcher::SpaceMatcher;
+use crate::rspace::trace::Log;
+use crate::rspace::trace::event::Produce;
+
+mod ispace_impl;
+mod locks;
+mod ops_consume;
+mod ops_install;
+mod ops_produce;
+mod setup;
+#[cfg(test)]
+mod tests;
+mod trace_log;
+
+pub use locks::LOCK_SEQUENCE;
 
 #[derive(Clone)]
 pub struct RSpaceStore {
@@ -69,33 +70,13 @@ pub struct RSpace<C, P, A, K> {
 
 impl<C, P, A, K> RSpace<C, P, A, K>
 where
-    C: Clone + Debug + Default + Serialize + std::hash::Hash + Ord + Eq + 'static + Sync + Send,
+    C: Clone + Debug + Default + Serialize + Hash + Ord + Eq + 'static + Sync + Send,
     P: Clone + Debug + Default + Serialize + 'static + Sync + Send,
     A: Clone + Debug + Default + Serialize + 'static + Sync + Send,
     K: Clone + Debug + Default + Serialize + 'static + Sync + Send,
 {
     pub fn get_store(&self) -> Arc<Box<dyn HotStore<C, P, A, K>>> {
         self.store.read().expect("store read lock").clone()
-    }
-
-    async fn consume_lock(&self, channel_hashes: &[u64]) -> (ChannelLockGuard, ChannelLockGuard) {
-        striped_locks::consume_lock(&self.phase_a_locks, &self.phase_b_locks, channel_hashes).await
-    }
-
-    async fn produce_lock(&self, channel: &C) -> (ChannelLockGuard, ChannelLockGuard) {
-        let channel_hash = striped_locks::channel_hash(channel);
-        let phase_a = striped_locks::acquire_locks(&self.phase_a_locks, &[channel_hash]).await;
-
-        let store = self.get_store();
-        let join_hashes: Vec<u64> = store
-            .get_joins(channel)
-            .into_iter()
-            .flatten()
-            .map(|ch| striped_locks::channel_hash(&ch))
-            .collect();
-
-        let phase_b = striped_locks::acquire_locks(&self.phase_b_locks, &join_hashes).await;
-        (phase_a, phase_b)
     }
 
     pub fn get_history_repository(
@@ -110,1235 +91,9 @@ where
 
 impl<C, P, A, K> SpaceMatcher<C, P, A, K> for RSpace<C, P, A, K>
 where
-    C: Clone + Debug + Default + Serialize + std::hash::Hash + Ord + Eq + 'static + Sync + Send,
-    P: Clone + Debug + Default + Serialize + 'static + Sync + Send,
-    A: Clone + Debug + Default + Serialize + 'static + Sync + Send,
-    K: Clone + Debug + Default + Serialize + 'static + Sync + Send,
-{
-}
-
-#[async_trait]
-impl<C, P, A, K> ISpace<C, P, A, K> for RSpace<C, P, A, K>
-where
-    C: Clone + Debug + Default + Serialize + std::hash::Hash + Ord + Eq + 'static + Sync + Send,
-    P: Clone + Debug + Default + Serialize + 'static + Sync + Send,
-    A: Clone + Debug + Default + Serialize + 'static + Sync + Send,
-    K: Clone + Debug + Default + Serialize + 'static + Sync + Send,
-{
-    async fn create_checkpoint(&self) -> Result<Checkpoint, RSpaceError> {
-        // Span[F].withMarks("create-checkpoint") from Scala - works because this is NOT
-        // async
-        let _span = tracing::info_span!(target: "f1r3fly.rspace", CREATE_CHECKPOINT_SPAN).entered();
-        tracing::trace!(target: "f1r3fly.rspace.ops", mark = "started-create-checkpoint", "create_checkpoint");
-
-        // Get changes with span
-        let changes = {
-            let _changes_span =
-                tracing::info_span!(target: "f1r3fly.rspace", CHANGES_SPAN).entered();
-            self.get_store().changes()
-        };
-
-        // Create history checkpoint with span
-        let next_history = {
-            let _history_span =
-                tracing::info_span!(target: "f1r3fly.rspace", HISTORY_CHECKPOINT_SPAN).entered();
-            self.get_history_repository().checkpoint(changes)
-        };
-        *self.history_repository.write().expect("history write lock") = Arc::new(next_history);
-
-        let log = std::mem::take(&mut *self.event_log.lock().expect("event log lock"));
-        let _ = std::mem::take(&mut *self.produce_counter.lock().expect("produce counter lock"));
-
-        let history_repo = self.get_history_repository();
-        let history_reader = history_repo.get_history_reader(&history_repo.root())?;
-
-        self.create_new_hot_store(history_reader);
-        self.restore_installs();
-
-        // Mark the completion of create-checkpoint
-        tracing::trace!(target: "f1r3fly.rspace.ops", mark = "finished-create-checkpoint", "create_checkpoint");
-
-        Ok(Checkpoint {
-            root: self.get_history_repository().root(),
-            log,
-        })
-    }
-
-    async fn reset(&self, root: &Blake2b256Hash) -> Result<(), RSpaceError> {
-        let _span = tracing::info_span!(target: "f1r3fly.rspace", RESET_SPAN).entered();
-        let next_history = self.get_history_repository().reset(root)?;
-        *self.history_repository.write().expect("history write lock") = Arc::new(next_history);
-
-        *self.event_log.lock().expect("event log lock") = Vec::new();
-        *self.produce_counter.lock().expect("produce counter lock") = BTreeMap::new();
-
-        // Striped locks are fixed-size and stateless (Mutex<()>); nothing to
-        // clear on reset — they are reused across checkpoints.
-
-        let history_reader = self.get_history_repository().get_history_reader(root)?;
-        self.create_new_hot_store(history_reader);
-        self.restore_installs();
-
-        Ok(())
-    }
-
-    async fn consume_result(
-        &self,
-        _channel: Vec<C>,
-        _pattern: Vec<P>,
-    ) -> Result<Option<(K, Vec<A>)>, RSpaceError> {
-        panic!("\nERROR: RSpace consume_result should not be called here");
-    }
-
-    async fn get_data(&self, channel: &C) -> Vec<Datum<A>> { self.get_store().get_data(channel) }
-
-    async fn get_waiting_continuations(&self, channels: Vec<C>) -> Vec<WaitingContinuation<P, K>> {
-        self.get_store().get_continuations(&channels)
-    }
-
-    async fn get_joins(&self, channel: C) -> Vec<Vec<C>> { self.get_store().get_joins(&channel) }
-
-    async fn clear(&self) -> Result<(), RSpaceError> {
-        self.reset(&RadixHistory::empty_root_node_hash()).await
-    }
-
-    async fn get_root(&self) -> Blake2b256Hash { self.get_history_repository().root() }
-
-    async fn to_map(&self) -> HashMap<Vec<C>, Row<P, A, K>> { self.get_store().to_map() }
-
-    async fn create_soft_checkpoint(&self) -> SoftCheckpoint<C, P, A, K> {
-        let cache_snapshot = self.get_store().snapshot();
-        let curr_event_log = std::mem::take(&mut *self.event_log.lock().expect("event log lock"));
-        let curr_produce_counter =
-            std::mem::take(&mut *self.produce_counter.lock().expect("produce counter lock"));
-
-        SoftCheckpoint {
-            cache_snapshot,
-            log: curr_event_log,
-            produce_counter: curr_produce_counter,
-        }
-    }
-
-    async fn take_event_log(&self) -> Log {
-        let curr_event_log = std::mem::take(&mut *self.event_log.lock().expect("event log lock"));
-        let _ = std::mem::take(&mut *self.produce_counter.lock().expect("produce counter lock"));
-        curr_event_log
-    }
-
-    async fn revert_to_soft_checkpoint(
-        &self,
-        checkpoint: SoftCheckpoint<C, P, A, K>,
-    ) -> Result<(), RSpaceError> {
-        let _span =
-            tracing::info_span!(target: "f1r3fly.rspace", REVERT_SOFT_CHECKPOINT_SPAN).entered();
-        let history = self.get_history_repository();
-        let history_reader = history.get_history_reader(&history.root())?;
-        let hot_store = HotStoreInstances::create_from_hs_and_hr(
-            checkpoint.cache_snapshot,
-            history_reader.base(),
-        );
-
-        *self.store.write().expect("store write lock") = Arc::new(hot_store);
-        *self.event_log.lock().expect("event log lock") = checkpoint.log;
-        *self.produce_counter.lock().expect("produce counter lock") = checkpoint.produce_counter;
-
-        Ok(())
-    }
-
-    async fn consume(
-        &self,
-        channels: Vec<C>,
-        patterns: Vec<P>,
-        continuation: K,
-        persist: bool,
-        peeks: BTreeSet<i32>,
-    ) -> Result<MaybeConsumeResult<C, P, A, K>, RSpaceError> {
-        if channels.is_empty() {
-            panic!("RUST ERROR: channels can't be empty");
-        }
-        if channels.len() != patterns.len() {
-            panic!("RUST ERROR: channels.length must equal patterns.length");
-        }
-        let consume_ref = Consume::create(&channels, &patterns, &continuation, persist);
-
-        let lock_start = Instant::now();
-        let channel_hashes: Vec<u64> = channels
-            .iter()
-            .map(|ch| striped_locks::channel_hash(ch))
-            .collect();
-        let _lock_guard = self.consume_lock(&channel_hashes).await;
-        if tracing::enabled!(target: "f1r3fly.rspace.lock_order", tracing::Level::TRACE) {
-            let seq = LOCK_SEQUENCE.fetch_add(1, AtomicOrdering::Relaxed);
-            tracing::trace!(target: "f1r3fly.rspace.lock_order", seq = seq, op = "consume", hashes = ?channel_hashes, "lock acquired");
-        }
-        metrics::counter!("rspace.consume.lock_acquire_ns", "source" => RSPACE_METRICS_SOURCE)
-            .increment(lock_start.elapsed().as_nanos() as u64);
-
-        metrics::counter!("rspace.consume.calls", "source" => RSPACE_METRICS_SOURCE).increment(1);
-        let start = Instant::now();
-        let result =
-            self.locked_consume(&channels, &patterns, &continuation, persist, &peeks, &consume_ref);
-        let duration = start.elapsed();
-        metrics::histogram!("comm_consume_time_seconds", "source" => RSPACE_METRICS_SOURCE)
-            .record(duration.as_secs_f64());
-        result
-    }
-
-    async fn produce(
-        &self,
-        channel: C,
-        data: A,
-        persist: bool,
-    ) -> Result<MaybeProduceResult<C, P, A, K>, RSpaceError> {
-        let produce_ref = Produce::create(&channel, &data, persist);
-
-        let lock_start = Instant::now();
-        let _lock_guard = self.produce_lock(&channel).await;
-        if tracing::enabled!(target: "f1r3fly.rspace.lock_order", tracing::Level::TRACE) {
-            let seq = LOCK_SEQUENCE.fetch_add(1, AtomicOrdering::Relaxed);
-            tracing::trace!(target: "f1r3fly.rspace.lock_order", seq = seq, op = "produce", hash = striped_locks::channel_hash(&channel), "lock acquired");
-        }
-        metrics::counter!("rspace.produce.lock_acquire_ns", "source" => RSPACE_METRICS_SOURCE)
-            .increment(lock_start.elapsed().as_nanos() as u64);
-
-        metrics::counter!("rspace.produce.calls", "source" => RSPACE_METRICS_SOURCE).increment(1);
-        let start = Instant::now();
-        let result = self.locked_produce(channel, data, persist, &produce_ref);
-        let duration = start.elapsed();
-        metrics::histogram!("comm_produce_time_seconds", "source" => RSPACE_METRICS_SOURCE)
-            .record(duration.as_secs_f64());
-        result
-    }
-
-    async fn install(
-        &self,
-        channels: Vec<C>,
-        patterns: Vec<P>,
-        continuation: K,
-    ) -> Result<Option<(K, Vec<A>)>, RSpaceError> {
-        metrics::counter!("rspace.install.calls", "source" => RSPACE_METRICS_SOURCE).increment(1);
-        let start = Instant::now();
-        let result = self.locked_install_internal(channels, patterns, continuation);
-        let duration = start.elapsed();
-        metrics::histogram!("install_time_seconds", "source" => RSPACE_METRICS_SOURCE)
-            .record(duration.as_secs_f64());
-        result
-    }
-
-    async fn rig_and_reset(
-        &self,
-        _start_root: Blake2b256Hash,
-        _log: Log,
-    ) -> Result<(), RSpaceError> {
-        panic!("\nERROR: RSpace rig_and_reset should not be called here");
-    }
-
-    async fn rig(&self, _log: Log) -> Result<(), RSpaceError> {
-        panic!("\nERROR: RSpace rig should not be called here");
-    }
-
-    async fn check_replay_data(&self) -> Result<(), RSpaceError> {
-        panic!("\nERROR: RSpace check_replay_data should not be called here");
-    }
-
-    async fn is_replay(&self) -> bool { false }
-
-    async fn update_produce(&self, produce_ref: Produce) -> () {
-        for event in self.event_log.lock().expect("event log lock").iter_mut() {
-            match event {
-                Event::IoEvent(IOEvent::Produce(produce)) => {
-                    if produce.hash == produce_ref.hash {
-                        *produce = produce_ref.clone();
-                    }
-                }
-
-                Event::Comm(comm) => {
-                    for p in comm.produces.iter_mut() {
-                        if p.hash == produce_ref.hash {
-                            *p = produce_ref.clone();
-                        }
-                    }
-                    if comm
-                        .times_repeated
-                        .keys()
-                        .any(|k| k.hash == produce_ref.hash)
-                    {
-                        comm.times_repeated = std::mem::take(&mut comm.times_repeated)
-                            .into_iter()
-                            .map(|(k, v)| {
-                                if k.hash == produce_ref.hash {
-                                    (produce_ref.clone(), v)
-                                } else {
-                                    (k, v)
-                                }
-                            })
-                            .collect();
-                    }
-                }
-
-                _ => continue,
-            }
-        }
-    }
-}
-
-impl<C, P, A, K> RSpace<C, P, A, K>
-where
     C: Clone + Debug + Default + Serialize + Hash + Ord + Eq + 'static + Sync + Send,
     P: Clone + Debug + Default + Serialize + 'static + Sync + Send,
     A: Clone + Debug + Default + Serialize + 'static + Sync + Send,
     K: Clone + Debug + Default + Serialize + 'static + Sync + Send,
 {
-    /**
-     * Creates [[RSpace]] from [[HistoryRepository]] and [[HotStore]].
-     */
-    pub fn apply(
-        history_repository: Arc<Box<dyn HistoryRepository<C, P, A, K> + Send + Sync + 'static>>,
-        store: Box<dyn HotStore<C, P, A, K>>,
-        matcher: Arc<Box<dyn Match<P, A, K>>>,
-    ) -> RSpace<C, P, A, K> {
-        RSpace {
-            history_repository: Arc::new(std::sync::RwLock::new(history_repository)),
-            store: Arc::new(std::sync::RwLock::new(Arc::new(store))),
-            matcher,
-            installs: Arc::new(std::sync::Mutex::new(HashMap::new())),
-            event_log: Arc::new(std::sync::Mutex::new(Vec::new())),
-            produce_counter: Arc::new(std::sync::Mutex::new(BTreeMap::new())),
-            phase_a_locks: striped_locks::new_striped_locks(),
-            phase_b_locks: striped_locks::new_striped_locks(),
-        }
-    }
-
-    pub fn create(
-        store: RSpaceStore,
-        matcher: Arc<Box<dyn Match<P, A, K>>>,
-    ) -> Result<RSpace<C, P, A, K>, HistoryRepositoryError>
-    where
-        C: Clone
-            + Debug
-            + Default
-            + Send
-            + Sync
-            + Serialize
-            + Ord
-            + Hash
-            + for<'a> Deserialize<'a>
-            + 'static,
-        P: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-        A: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-        K: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-    {
-        let setup = Self::create_history_repo(store).unwrap();
-        let (history_reader, store) = setup;
-        let space = Self::apply(Arc::new(history_reader), store, matcher);
-        Ok(space)
-    }
-
-    pub fn create_with_replay(
-        store: RSpaceStore,
-        matcher: Arc<Box<dyn Match<P, A, K>>>,
-    ) -> Result<(RSpace<C, P, A, K>, ReplayRSpace<C, P, A, K>), HistoryRepositoryError>
-    where
-        C: Clone
-            + Debug
-            + Default
-            + Send
-            + Sync
-            + Serialize
-            + Ord
-            + Hash
-            + for<'a> Deserialize<'a>
-            + 'static,
-        P: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-        A: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-        K: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-    {
-        let setup = Self::create_history_repo(store).unwrap();
-        let (history_repo, store) = setup;
-        let history_repo_arc = Arc::new(history_repo);
-
-        // Play
-        let space = Self::apply(history_repo_arc.clone(), store, matcher.clone());
-        // Replay
-        let history_reader: Box<dyn HistoryReader<Blake2b256Hash, C, P, A, K>> =
-            history_repo_arc.get_history_reader(&history_repo_arc.root())?;
-        let replay_store = HotStoreInstances::create_from_hr(history_reader.base());
-        let replay = ReplayRSpace::apply_with_logger(
-            history_repo_arc.clone(),
-            Arc::new(replay_store),
-            matcher.clone(),
-            Box::new(BasicLogger::new()),
-        );
-        Ok((space, replay))
-    }
-
-    /**
-     * Creates [[HistoryRepository]] and [[HotStore]].
-     */
-    pub fn create_history_repo(
-        store: RSpaceStore,
-    ) -> Result<
-        (
-            Box<dyn HistoryRepository<C, P, A, K> + Send + Sync + 'static>,
-            Box<dyn HotStore<C, P, A, K>>,
-        ),
-        HistoryRepositoryError,
-    >
-    where
-        C: Clone
-            + Debug
-            + Default
-            + Send
-            + Sync
-            + Serialize
-            + for<'a> Deserialize<'a>
-            + Eq
-            + Hash
-            + 'static,
-        P: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-        A: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-        K: Clone + Debug + Default + Send + Sync + Serialize + for<'a> Deserialize<'a> + 'static,
-    {
-        let history_repo =
-            HistoryRepositoryInstances::lmdb_repository(store.history, store.roots, store.cold)?;
-
-        let history_reader = history_repo.get_history_reader(&history_repo.root())?;
-
-        let hot_store = HotStoreInstances::create_from_hr(history_reader.base());
-
-        Ok((history_repo, hot_store))
-    }
-
-    fn produce_counters(&self, produce_refs: &[Produce]) -> BTreeMap<Produce, i32> {
-        let counter = self.produce_counter.lock().expect("produce counter lock");
-        produce_refs
-            .iter()
-            .map(|p| (p.clone(), counter.get(p).copied().unwrap_or(0)))
-            .collect()
-    }
-
-    fn locked_consume(
-        &self,
-        channels: &[C],
-        patterns: &[P],
-        continuation: &K,
-        persist: bool,
-        peeks: &BTreeSet<i32>,
-        consume_ref: &Consume,
-    ) -> Result<MaybeConsumeResult<C, P, A, K>, RSpaceError> {
-        // Span[F].traceI("locked-consume") from Scala
-        let _span = tracing::info_span!(target: "f1r3fly.rspace", LOCKED_CONSUME_SPAN).entered();
-        tracing::trace!(target: "f1r3fly.rspace.ops", mark = "started-locked-consume", "locked_consume");
-
-        let t0 = Instant::now();
-        self.log_consume(consume_ref);
-        metrics::counter!("rspace.consume.log_ns", "source" => RSPACE_METRICS_SOURCE)
-            .increment(t0.elapsed().as_nanos() as u64);
-
-        let t1 = Instant::now();
-        let mut channel_to_indexed_data = self.fetch_channel_to_index_data(channels);
-        metrics::counter!("rspace.consume.fetch_data_ns", "source" => RSPACE_METRICS_SOURCE)
-            .increment(t1.elapsed().as_nanos() as u64);
-
-        let t2 = Instant::now();
-        let zipped: Vec<(C, P)> = channels
-            .iter()
-            .cloned()
-            .zip(patterns.iter().cloned())
-            .collect();
-        let options: Option<Vec<ConsumeCandidate<C, A>>> = self
-            .extract_data_candidates(&self.matcher, &zipped, &mut channel_to_indexed_data)
-            .into_iter()
-            .collect();
-        metrics::counter!("rspace.consume.match_ns", "source" => RSPACE_METRICS_SOURCE)
-            .increment(t2.elapsed().as_nanos() as u64);
-
-        let wk = WaitingContinuation {
-            patterns: patterns.to_vec(),
-            continuation: continuation.clone(),
-            persist,
-            peeks: peeks.clone(),
-            source: consume_ref.clone(),
-        };
-
-        let commit_ok = match &options {
-            // Cross-channel commit hook: a `where` guard on the consume
-            // can veto even after every spatial bind matched. On false
-            // we install the wk and leave the data alone, just as if
-            // the spatial match itself had failed (plan §7.12).
-            Some(data_candidates) => {
-                let matched: Vec<A> = data_candidates.iter().map(|c| c.datum.a.clone()).collect();
-                self.matcher.check_commit(continuation, &matched)
-            }
-            None => false,
-        };
-
-        match options {
-            Some(data_candidates) if commit_ok => {
-                let t3 = Instant::now();
-                let produce_counters_closure =
-                    |produces: &[Produce]| self.produce_counters(produces);
-
-                self.log_comm(
-                    COMM::new(
-                        &data_candidates,
-                        consume_ref.clone(),
-                        peeks.clone(),
-                        produce_counters_closure,
-                    ),
-                    CONSUME_COMM_LABEL,
-                );
-                self.store_persistent_data(&data_candidates);
-                metrics::counter!("rspace.consume.process_match_ns", "source" => RSPACE_METRICS_SOURCE)
-                    .increment(t3.elapsed().as_nanos() as u64);
-                tracing::trace!(target: "f1r3fly.rspace.ops", mark = "finished-locked-consume", "locked_consume");
-                Ok(self.wrap_result(channels, &wk, &data_candidates))
-            }
-            _ => {
-                let t3 = Instant::now();
-                self.store_waiting_continuation(channels.to_vec(), wk);
-                metrics::counter!("rspace.consume.store_continuation_ns", "source" => RSPACE_METRICS_SOURCE)
-                    .increment(t3.elapsed().as_nanos() as u64);
-                tracing::trace!(target: "f1r3fly.rspace.ops", mark = "finished-locked-consume", "locked_consume");
-                Ok(None)
-            }
-        }
-    }
-
-    /*
-     * Here, we create a cache of the data at each channel as
-     * `channelToIndexedData` which is used for finding matches.  When a
-     * speculative match is found, we can remove the matching datum from the
-     * remaining data candidates in the cache.
-     *
-     * Put another way, this allows us to speculatively remove matching data
-     * without affecting the actual store contents.
-     */
-    fn fetch_channel_to_index_data(&self, channels: &[C]) -> HashMap<C, Vec<(Datum<A>, i32)>> {
-        let store = self.get_store();
-        let mut map = HashMap::with_capacity(channels.len());
-        for c in channels {
-            let data = store.get_data(c);
-            let shuffled_data = self.shuffle_with_index(data);
-            map.insert(c.clone(), shuffled_data);
-        }
-        map
-    }
-
-    fn locked_produce(
-        &self,
-        channel: C,
-        data: A,
-        persist: bool,
-        produce_ref: &Produce,
-    ) -> Result<MaybeProduceResult<C, P, A, K>, RSpaceError> {
-        // Span[F].traceI("locked-produce") from Scala
-        let _span = tracing::info_span!(target: "f1r3fly.rspace", LOCKED_PRODUCE_SPAN).entered();
-        tracing::trace!(target: "f1r3fly.rspace.ops", mark = "started-locked-produce", "locked_produce");
-
-        let t0 = Instant::now();
-        let grouped_channels = self.get_store().get_joins(&channel);
-        metrics::counter!("rspace.produce.get_joins_ns", "source" => RSPACE_METRICS_SOURCE)
-            .increment(t0.elapsed().as_nanos() as u64);
-
-        self.log_produce(produce_ref, persist);
-
-        let t1 = Instant::now();
-        let extracted = self.extract_produce_candidate(grouped_channels, channel.clone(), Datum {
-            a: data.clone(),
-            persist,
-            source: produce_ref.clone(),
-        });
-        metrics::counter!("rspace.produce.extract_candidate_ns", "source" => RSPACE_METRICS_SOURCE)
-            .increment(t1.elapsed().as_nanos() as u64);
-
-        match extracted {
-            Some(produce_candidate) => {
-                let t2 = Instant::now();
-                let result =
-                    Ok(self
-                        .process_match_found(produce_candidate)
-                        .map(|consume_result| {
-                            (consume_result.0, consume_result.1, produce_ref.clone())
-                        }));
-                metrics::counter!("rspace.produce.process_match_ns", "source" => RSPACE_METRICS_SOURCE)
-                    .increment(t2.elapsed().as_nanos() as u64);
-                tracing::trace!(target: "f1r3fly.rspace.ops", mark = "finished-locked-produce", "locked_produce");
-                result
-            }
-            None => {
-                let t2 = Instant::now();
-                let result = Ok(self.store_data(channel, data, persist, produce_ref.clone()));
-                metrics::counter!("rspace.produce.store_data_ns", "source" => RSPACE_METRICS_SOURCE)
-                    .increment(t2.elapsed().as_nanos() as u64);
-                tracing::trace!(target: "f1r3fly.rspace.ops", mark = "finished-locked-produce", "locked_produce");
-                result
-            }
-        }
-    }
-
-    /*
-     * Find produce candidate
-     *
-     * NOTE: On Rust side, we are NOT passing functions through. Instead just the
-     * data. And then in 'run_matcher_for_channels' we call the functions
-     * defined below
-     */
-    fn extract_produce_candidate(
-        &self,
-        grouped_channels: Vec<Vec<C>>,
-        bat_channel: C,
-        data: Datum<A>,
-    ) -> MaybeProduceCandidate<C, P, A, K> {
-        let fetch_matching_continuations =
-            |channels: Vec<C>| -> Vec<(std::sync::Arc<WaitingContinuation<P, K>>, i32)> {
-                // Arc-shared fetch: probing continuations no longer deep-clones
-                // the continuation body on every produce.
-                let continuations = self.get_store().get_continuations_arc(&channels);
-                self.shuffle_with_index(continuations)
-            };
-
-        /*
-         * Here, we create a cache of the data at each channel as
-         * `channelToIndexedData` which is used for finding matches.  When a
-         * speculative match is found, we can remove the matching datum from
-         * the remaining data candidates in the cache.
-         *
-         * Put another way, this allows us to speculatively remove matching data
-         * without affecting the actual store contents.
-         *
-         * In this version, we also add the produced data directly to this cache.
-         */
-        let fetch_matching_data = |channel| -> (C, Vec<(Datum<A>, i32)>) {
-            let data_vec = self.get_store().get_data(&channel);
-            let mut shuffled_data = self.shuffle_with_index(data_vec);
-            if channel == bat_channel {
-                shuffled_data.insert(0, (data.clone(), -1));
-            }
-            (channel, shuffled_data)
-        };
-
-        self.run_matcher_for_channels(
-            grouped_channels,
-            fetch_matching_continuations,
-            fetch_matching_data,
-        )
-    }
-
-    fn process_match_found(
-        &self,
-        pc: ProduceCandidate<C, P, A, K>,
-    ) -> MaybeConsumeResult<C, P, A, K> {
-        let ProduceCandidate {
-            channels,
-            continuation,
-            continuation_index,
-            data_candidates,
-        } = pc;
-
-        let WaitingContinuation {
-            patterns: _patterns,
-            continuation: _cont,
-            persist,
-            peeks,
-            source: consume_ref,
-        } = &continuation;
-
-        let produce_counters_closure = |produces: &[Produce]| self.produce_counters(produces);
-        self.log_comm(
-            COMM::new(
-                &data_candidates,
-                consume_ref.clone(),
-                peeks.clone(),
-                produce_counters_closure,
-            ),
-            PRODUCE_COMM_LABEL,
-        );
-
-        if !persist {
-            self.get_store()
-                .remove_continuation(&channels, continuation_index);
-        }
-
-        self.remove_matched_datum_and_join(&channels, &data_candidates);
-
-        self.wrap_result(&channels, &continuation, &data_candidates)
-    }
-
-    fn log_comm(&self, comm: COMM, comm_metric_label: &'static str) {
-        // Increment counter FIRST (matching Scala) using constants to avoid
-        // memory leaks. The label is CONSUME_COMM_LABEL or PRODUCE_COMM_LABEL.
-        metrics::counter!(comm_metric_label, "source" => RSPACE_METRICS_SOURCE).increment(1);
-
-        // Then update event log (RSpace-specific behavior)
-        self.event_log
-            .lock()
-            .expect("event log lock")
-            .push(Event::Comm(comm));
-    }
-
-    fn log_consume(&self, consume_ref: &Consume) {
-        self.event_log
-            .lock()
-            .expect("event log lock")
-            .push(Event::IoEvent(IOEvent::Consume(consume_ref.clone())));
-    }
-
-    fn log_produce(&self, produce_ref: &Produce, persist: bool) {
-        self.event_log
-            .lock()
-            .expect("event log lock")
-            .push(Event::IoEvent(IOEvent::Produce(produce_ref.clone())));
-        if !persist {
-            let mut counter = self.produce_counter.lock().expect("produce counter lock");
-            match counter.get_mut(produce_ref) {
-                Some(count) => *count += 1,
-                None => {
-                    counter.insert(produce_ref.clone(), 1);
-                }
-            }
-        }
-    }
-
-    pub fn spawn(&self) -> Result<Self, RSpaceError> {
-        // Span[F].withMarks("spawn") from Scala - works because this is NOT async
-        let _span = tracing::info_span!(target: "f1r3fly.rspace", "spawn").entered();
-        tracing::trace!(target: "f1r3fly.rspace.ops", mark = "started-spawn", "spawn");
-
-        let history_repo = self.get_history_repository();
-        let next_history = history_repo.reset(&history_repo.root())?;
-        let history_reader = next_history.get_history_reader(&next_history.root())?;
-        let hot_store = HotStoreInstances::create_from_hr(history_reader.base());
-        let rspace = RSpace::apply(Arc::new(next_history), hot_store, self.matcher.clone());
-        rspace.restore_installs();
-
-        // Mark the completion of spawn operation
-        tracing::trace!(target: "f1r3fly.rspace.ops", mark = "finished-spawn", "spawn");
-        Ok(rspace)
-    }
-
-    /* RSpaceOps */
-
-    fn store_waiting_continuation(
-        &self,
-        channels: Vec<C>,
-        wc: WaitingContinuation<P, K>,
-    ) -> MaybeConsumeResult<C, P, A, K> {
-        let store = self.get_store();
-        let _ = store.put_continuation(&channels, wc);
-        for channel in channels.iter() {
-            store.put_join(channel, &channels);
-        }
-        None
-    }
-
-    fn store_data(
-        &self,
-        channel: C,
-        data: A,
-        persist: bool,
-        produce_ref: Produce,
-    ) -> MaybeProduceResult<C, P, A, K> {
-        self.get_store().put_datum(&channel, Datum {
-            a: data,
-            persist,
-            source: produce_ref,
-        });
-
-        None
-    }
-
-    fn store_persistent_data(&self, data_candidates: &[ConsumeCandidate<C, A>]) {
-        let mut sorted_candidates: Vec<_> = data_candidates.iter().collect();
-        sorted_candidates.sort_by_key(|candidate| candidate.datum_index);
-        let store = self.get_store();
-        for consume_candidate in sorted_candidates {
-            if !consume_candidate.datum.persist {
-                let _ =
-                    store.remove_datum(&consume_candidate.channel, consume_candidate.datum_index);
-            }
-        }
-    }
-
-    fn restore_installs(&self) {
-        // Move out the install map to avoid cloning the whole structure on each
-        // restore.
-        let installs = std::mem::take(&mut *self.installs.lock().unwrap());
-        for (channels, install) in installs {
-            self.locked_install_internal(channels, install.patterns, install.continuation)
-                .unwrap();
-        }
-    }
-
-    fn locked_install_internal(
-        &self,
-        channels: Vec<C>,
-        patterns: Vec<P>,
-        continuation: K,
-    ) -> Result<Option<(K, Vec<A>)>, RSpaceError> {
-        if channels.len() != patterns.len() {
-            panic!("RUST ERROR: channels.length must equal patterns.length");
-        }
-        let consume_ref = Consume::create(&channels, &patterns, &continuation, true);
-        let mut channel_to_indexed_data = self.fetch_channel_to_index_data(&channels);
-        let zipped: Vec<(C, P)> = channels
-            .iter()
-            .cloned()
-            .zip(patterns.iter().cloned())
-            .collect();
-        let options: Option<Vec<ConsumeCandidate<C, A>>> = self
-            .extract_data_candidates(&self.matcher, &zipped, &mut channel_to_indexed_data)
-            .into_iter()
-            .collect();
-
-        match options {
-            None => {
-                self.installs
-                    .lock()
-                    .unwrap()
-                    .insert(channels.clone(), Install {
-                        patterns: patterns.clone(),
-                        continuation: continuation.clone(),
-                    });
-
-                let store = self.get_store();
-                store.install_continuation(&channels, WaitingContinuation {
-                    patterns,
-                    continuation,
-                    persist: true,
-                    peeks: BTreeSet::default(),
-                    source: consume_ref,
-                });
-
-                for channel in channels.iter() {
-                    store.install_join(channel, &channels);
-                }
-                Ok(None)
-            }
-            Some(_) => Err(RSpaceError::BugFoundError(
-                "RUST ERROR: Installing can be done only on startup".to_string(),
-            )),
-        }
-    }
-
-    fn create_new_hot_store(
-        &self,
-        history_reader: Box<dyn HistoryReader<Blake2b256Hash, C, P, A, K>>,
-    ) {
-        let next_hot_store = HotStoreInstances::create_from_hr(history_reader.base());
-        *self.store.write().expect("store write lock") = Arc::new(next_hot_store);
-    }
-
-    fn wrap_result(
-        &self,
-        channels: &[C],
-        wk: &WaitingContinuation<P, K>,
-        data_candidates: &[ConsumeCandidate<C, A>],
-    ) -> MaybeConsumeResult<C, P, A, K> {
-        let cont_result = ContResult {
-            continuation: wk.continuation.clone(),
-            persistent: wk.persist,
-            channels: channels.to_vec(),
-            patterns: wk.patterns.clone(),
-            peek: !wk.peeks.is_empty(),
-        };
-
-        let rspace_results = data_candidates
-            .iter()
-            .map(|data_candidate| RSpaceResult {
-                channel: data_candidate.channel.clone(),
-                matched_datum: data_candidate.datum.a.clone(),
-                removed_datum: data_candidate.removed_datum.clone(),
-                persistent: data_candidate.datum.persist,
-            })
-            .collect();
-
-        Some((cont_result, rspace_results))
-    }
-
-    fn remove_matched_datum_and_join(
-        &self,
-        channels: &[C],
-        data_candidates: &[ConsumeCandidate<C, A>],
-    ) {
-        let mut sorted_candidates: Vec<_> = data_candidates.iter().collect();
-        sorted_candidates.sort_by_key(|candidate| candidate.datum_index);
-        let store = self.get_store();
-        for consume_candidate in sorted_candidates {
-            let channel = &consume_candidate.channel;
-            let datum_index = consume_candidate.datum_index;
-
-            if datum_index >= 0 &&
-                !consume_candidate.datum.persist &&
-                store.remove_datum(channel, datum_index).is_err()
-            {
-                continue;
-            }
-            store.remove_join(channel, channels);
-        }
-    }
-
-    fn run_matcher_for_channels(
-        &self,
-        grouped_channels: Vec<Vec<C>>,
-        fetch_matching_continuations: impl Fn(
-            Vec<C>,
-        )
-            -> Vec<(std::sync::Arc<WaitingContinuation<P, K>>, i32)>,
-        fetch_matching_data: impl Fn(C) -> (C, Vec<(Datum<A>, i32)>),
-    ) -> MaybeProduceCandidate<C, P, A, K> {
-        for channels in &grouped_channels {
-            let t_cont = Instant::now();
-            let match_candidates = fetch_matching_continuations(channels.to_vec());
-            metrics::counter!("rspace.matcher.fetch_continuations_ns", "source" => RSPACE_METRICS_SOURCE)
-                .increment(t_cont.elapsed().as_nanos() as u64);
-            metrics::counter!("rspace.matcher.continuations_returned", "source" => RSPACE_METRICS_SOURCE)
-                .increment(match_candidates.len() as u64);
-
-            let t_data = Instant::now();
-            let channel_to_indexed_data: HashMap<C, Vec<(Datum<A>, i32)>> = channels
-                .iter()
-                .map(|c| fetch_matching_data(c.clone()))
-                .collect();
-            metrics::counter!("rspace.matcher.fetch_data_ns", "source" => RSPACE_METRICS_SOURCE)
-                .increment(t_data.elapsed().as_nanos() as u64);
-
-            let t_match = Instant::now();
-            let first_match = self.extract_first_match(
-                &self.matcher,
-                channels.to_vec(),
-                match_candidates,
-                channel_to_indexed_data,
-            );
-            metrics::counter!("rspace.matcher.extract_first_match_ns", "source" => RSPACE_METRICS_SOURCE)
-                .increment(t_match.elapsed().as_nanos() as u64);
-
-            if let Some(produce_candidate) = first_match {
-                return Some(produce_candidate);
-            }
-        }
-        None
-    }
-
-    fn shuffle_with_index<D>(&self, t: Vec<D>) -> Vec<(D, i32)> {
-        let mut indexed_vec = t
-            .into_iter()
-            .enumerate()
-            .map(|(i, d)| (d, i as i32))
-            .collect::<Vec<_>>();
-        if indexed_vec.len() >= 2 {
-            indexed_vec.shuffle(&mut rand::rng());
-        }
-        indexed_vec
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use std::sync::Arc;
-    use std::sync::atomic::{AtomicBool, AtomicU64, Ordering};
-
-    use serde::{Deserialize, Serialize};
-
-    use super::*;
-    use crate::rspace::r#match::Match;
-    use crate::rspace::shared::in_mem_store_manager::InMemoryStoreManager;
-    use crate::rspace::shared::key_value_store_manager::KeyValueStoreManager;
-
-    // ── minimal types ─────────────────────────────────────────────────────────
-
-    #[derive(Clone, Debug, Serialize, Deserialize, Default, PartialEq)]
-    struct Wildcard;
-
-    #[derive(Clone, Debug, Serialize, Deserialize, Default, PartialEq)]
-    struct Cont;
-
-    struct AlwaysMatch;
-
-    impl Match<Wildcard, String, Cont> for AlwaysMatch {
-        fn get(&self, _: &Wildcard, a: &String) -> Option<String> { Some(a.clone()) }
-    }
-
-    async fn make_rspace() -> RSpace<String, Wildcard, String, Cont> {
-        let mut kvm = InMemoryStoreManager::new();
-        let store = kvm.r_space_stores().await.unwrap();
-        RSpace::create(store, Arc::new(Box::new(AlwaysMatch))).unwrap()
-    }
-
-    // Measures contention on the event_log mutex while N concurrent tasks call
-    // produce() on separate channels. The log is pre-filled to PRE_FILL entries
-    // before the concurrent phase so every insert starts with a large existing
-    // log, making the mutex hold time long enough to observe.
-    //
-    // Observer runs on a dedicated OS thread (not a tokio task) because
-    // std::sync::Mutex::lock() blocks the worker thread without yielding, so a
-    // tokio observer would never be scheduled while producers hold the mutex.
-    //
-    // Passes when event_log uses O(1) append: hold time drops to ~10 ns and
-    // the observer almost never catches the mutex held.
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-    async fn event_log_mutex_does_not_contend_under_concurrent_produces() {
-        const TASKS: usize = 4;
-        const OPS_PER_TASK: usize = 200;
-        // Simulates mid-deploy state: the event_log grows to PRE_FILL entries
-        // before the concurrent test starts. Each subsequent Vec::insert(0,..)
-        // must shift all existing entries — O(PRE_FILL × sizeof(Event)) bytes.
-        // At PRE_FILL=8000 and sizeof(Event)≈150 bytes on M1:
-        //   shift ≈ 1.2 MB / 50 GB/s ≈ 24 μs per insert → detectable.
-        const PRE_FILL: usize = 8_000;
-        // Fraction of observer probes that find the mutex already held.
-        const MAX_CONTENTION_RATE: f64 = 0.20;
-
-        let rspace = make_rspace().await;
-
-        // Observer on a dedicated OS thread: probes try_lock() in a spin loop.
-        // Must be an OS thread, not a tokio task — std::sync::Mutex::lock()
-        // blocks the worker thread without yielding, so a tokio observer would
-        // never run while any producer holds the mutex.
-        let event_log = rspace.event_log.clone();
-        let running = Arc::new(AtomicBool::new(true));
-        let total_probes = Arc::new(AtomicU64::new(0));
-        let contended_probes = Arc::new(AtomicU64::new(0));
-
-        {
-            let event_log = event_log.clone();
-            let running = running.clone();
-            let total = total_probes.clone();
-            let contended = contended_probes.clone();
-            std::thread::spawn(move || {
-                while running.load(Ordering::Relaxed) {
-                    total.fetch_add(1, Ordering::Relaxed);
-                    if event_log.try_lock().is_err() {
-                        contended.fetch_add(1, Ordering::Relaxed);
-                    }
-                    std::hint::spin_loop();
-                }
-            });
-        }
-
-        // Pre-fill: grow the log to PRE_FILL entries before the concurrent phase.
-        // Counters are reset after so we measure only concurrent contention.
-        for i in 0..PRE_FILL {
-            rspace
-                .produce(format!("prefill_{}", i), "datum".to_string(), false)
-                .await
-                .unwrap();
-        }
-
-        total_probes.store(0, Ordering::Relaxed);
-        contended_probes.store(0, Ordering::Relaxed);
-
-        // N concurrent producers, each on its own channel set.
-        let handles: Vec<_> = (0..TASKS)
-            .map(|i| {
-                let s = rspace.clone();
-                tokio::spawn(async move {
-                    for j in 0..OPS_PER_TASK {
-                        s.produce(format!("ch_{}_{}", i, j), "datum".to_string(), false)
-                            .await
-                            .unwrap();
-                    }
-                })
-            })
-            .collect();
-        for h in handles {
-            h.await.unwrap();
-        }
-        running.store(false, Ordering::Relaxed);
-
-        let total = total_probes.load(Ordering::Relaxed);
-        let contended = contended_probes.load(Ordering::Relaxed);
-        let rate = if total > 0 {
-            contended as f64 / total as f64
-        } else {
-            0.0
-        };
-
-        eprintln!(
-            "event_log_contention: probes={total}  contended={contended}  rate={:.1}%  (threshold \
-             <{:.0}%)",
-            rate * 100.0,
-            MAX_CONTENTION_RATE * 100.0,
-        );
-
-        assert!(
-            rate < MAX_CONTENTION_RATE,
-            "event_log mutex contention too high: {:.1}% of probes found the mutex held \
-             (threshold {:.0}%). Root cause: all {} concurrent tasks share one \
-             std::sync::Mutex<event_log> — every produce() call acquires it, blocking other \
-             worker threads. Fix: per-task event logs merged at checkpoint, or a lock-free append \
-             structure.",
-            rate * 100.0,
-            MAX_CONTENTION_RATE * 100.0,
-            TASKS,
-        );
-    }
-
-    // Mirrors the rholang-par benchmark: PAR_BRANCHES concurrent tokio tasks
-    // each call produce() OPS_PER_BRANCH times on their own private channels,
-    // all sharing one RSpace and therefore one event_log.
-    //
-    // The log grows naturally from zero to PAR_BRANCHES * OPS_PER_BRANCH entries.
-    // Passes when event_log uses O(1) append.
-
-    #[tokio::test(flavor = "multi_thread", worker_threads = 8)]
-    async fn par_branch_event_log_does_not_contend_at_rholang_par_scale() {
-        const PAR_BRANCHES: usize = 32;
-        const OPS_PER_BRANCH: usize = 500;
-        const MAX_CONTENTION_RATE: f64 = 0.20;
-
-        let rspace = make_rspace().await;
-
-        let event_log = rspace.event_log.clone();
-        let running = Arc::new(AtomicBool::new(true));
-        let total_probes = Arc::new(AtomicU64::new(0));
-        let contended_probes = Arc::new(AtomicU64::new(0));
-
-        {
-            let event_log = event_log.clone();
-            let running = running.clone();
-            let total = total_probes.clone();
-            let contended = contended_probes.clone();
-            std::thread::spawn(move || {
-                while running.load(Ordering::Relaxed) {
-                    total.fetch_add(1, Ordering::Relaxed);
-                    if event_log.try_lock().is_err() {
-                        contended.fetch_add(1, Ordering::Relaxed);
-                    }
-                    std::hint::spin_loop();
-                }
-            });
-        }
-
-        // 32 par-branches, each producing on its own unique channels.
-        // No matching happens, so contention comes purely from log growth.
-        let handles: Vec<_> = (0..PAR_BRANCHES)
-            .map(|i| {
-                let s = rspace.clone();
-                tokio::spawn(async move {
-                    for j in 0..OPS_PER_BRANCH {
-                        s.produce(format!("branch_{}_{}", i, j), "datum".to_string(), false)
-                            .await
-                            .unwrap();
-                    }
-                })
-            })
-            .collect();
-        for h in handles {
-            h.await.unwrap();
-        }
-        running.store(false, Ordering::Relaxed);
-
-        let total = total_probes.load(Ordering::Relaxed);
-        let contended = contended_probes.load(Ordering::Relaxed);
-        let rate = if total > 0 {
-            contended as f64 / total as f64
-        } else {
-            0.0
-        };
-
-        eprintln!(
-            "par_branch_contention: branches={PAR_BRANCHES}  ops_per_branch={OPS_PER_BRANCH}  \
-             total_ops={}  probes={total}  contended={contended}  rate={:.1}%  (threshold <{:.0}%)",
-            PAR_BRANCHES * OPS_PER_BRANCH,
-            rate * 100.0,
-            MAX_CONTENTION_RATE * 100.0,
-        );
-
-        assert!(
-            rate < MAX_CONTENTION_RATE,
-            "event_log mutex contention too high at {PAR_BRANCHES} par-branches: {:.1}% of probes \
-             found the mutex held (threshold {:.0}%). Root cause: all {PAR_BRANCHES} par-branch \
-             tasks share one std::sync::Mutex<event_log> and each produce() calls \
-             Vec::insert(0,..) — O(n) shift where n grows to {} entries. Fix: per-branch event \
-             logs merged at create_checkpoint, or replace insert(0,..) with push().",
-            rate * 100.0,
-            MAX_CONTENTION_RATE * 100.0,
-            PAR_BRANCHES * OPS_PER_BRANCH,
-        );
-    }
-
-    // Disambiguates two candidate explanations for the near-zero wall-clock
-    // speedup measured end-to-end at rholang-par scale (bench_par_branches:
-    // 0.88-0.96x at 32 branches against a worker-thread-bounded ideal — see
-    // issue #50 follow-up):
-    // (a) event_log/produce_counter themselves cost that much, or
-    // (b) something else in produce()'s path (HotStore, the striped
-    //     per-channel lock, tokio scheduling) is the real cost and these
-    //     two locks are not the story despite being std::sync::Mutex.
-    //
-    // Calls log_produce() directly — the exact call log_produce() itself
-    // takes (event_log.push then, for non-persist, produce_counter.insert)
-    // with no produce_lock(), no HotStore, no get_store() RwLock read, no
-    // matcher involved. Isolates these two locks from every other cost
-    // produce() incurs, at the same branch/op counts as bench_par_branches.
-    #[tokio::test(flavor = "multi_thread", worker_threads = 8)]
-    async fn event_log_and_produce_counter_isolated_cost_at_rholang_par_scale() {
-        const PAR_BRANCHES: usize = 32;
-        const OPS_PER_BRANCH: usize = 5000;
-
-        // Sequential baseline: identical total ops, one after another.
-        let seq_rspace = make_rspace().await;
-        let t_seq = Instant::now();
-        for b in 0..PAR_BRANCHES {
-            for i in 0..OPS_PER_BRANCH {
-                let channel = format!("seq_{}_{}", b, i);
-                let data = "datum".to_string();
-                let produce_ref = Produce::create(&channel, &data, false);
-                seq_rspace.log_produce(&produce_ref, false);
-            }
-        }
-        let seq_ms = t_seq.elapsed().as_millis().max(1);
-
-        // Concurrent: PAR_BRANCHES tokio tasks, each on its own channel set,
-        // all sharing one RSpace and therefore one event_log/produce_counter.
-        let par_rspace = Arc::new(make_rspace().await);
-        let t_par = Instant::now();
-        let handles: Vec<_> = (0..PAR_BRANCHES)
-            .map(|b| {
-                let s = par_rspace.clone();
-                tokio::spawn(async move {
-                    for i in 0..OPS_PER_BRANCH {
-                        let channel = format!("par_{}_{}", b, i);
-                        let data = "datum".to_string();
-                        let produce_ref = Produce::create(&channel, &data, false);
-                        s.log_produce(&produce_ref, false);
-                    }
-                })
-            })
-            .collect();
-        for h in handles {
-            h.await.unwrap();
-        }
-        let par_ms = t_par.elapsed().as_millis().max(1);
-
-        // log_produce() is synchronous and the branch loops never await, so
-        // each task holds its worker for the whole loop: at most num_workers
-        // (further capped by the host's cores) branches run at once. That, not
-        // PAR_BRANCHES, is the achievable ideal for the efficiency metric.
-        let num_workers = tokio::runtime::Handle::current().metrics().num_workers();
-        let ideal = PAR_BRANCHES.min(num_workers).min(
-            std::thread::available_parallelism()
-                .map(|n| n.get())
-                .unwrap_or(usize::MAX),
-        );
-        let speedup = seq_ms as f64 / par_ms as f64;
-        let efficiency = speedup / ideal as f64 * 100.0;
-
-        eprintln!(
-            "event_log+produce_counter isolated cost: branches={PAR_BRANCHES} \
-             ops_per_branch={OPS_PER_BRANCH} total_ops={} sequential={seq_ms}ms \
-             concurrent={par_ms}ms speedup={:.2}x (ideal {ideal}x: {PAR_BRANCHES} branches on \
-             {num_workers} workers) efficiency={:.1}%",
-            PAR_BRANCHES * OPS_PER_BRANCH,
-            speedup,
-            efficiency,
-        );
-
-        // No assertion: this test is diagnostic, not a regression gate. It
-        // reports the isolated cost of these two locks so it can be compared
-        // against bench_par_branches' end-to-end number for the same
-        // branch/op counts (see issue #50 investigation).
-    }
 }
